+++ tests/unit/test_models/test_full_battery_models/test_base_battery_model.py
@@ -434,12 +434,12 @@ class TestBaseBatteryModel:
 
     def test_default_solver(self):
         model = pybamm.BaseBatteryModel()
-        assert isinstance(model.default_solver, pybamm.CasadiSolver)
+        assert isinstance(model.default_solver, pybamm.IDAKLUSolver)
 
         # check that default_solver gives you a new solver, not an internal object
         solver = model.default_solver
         solver = pybamm.BaseModel()
-        assert isinstance(model.default_solver, pybamm.CasadiSolver)
+        assert isinstance(model.default_solver, pybamm.IDAKLUSolver)
         assert isinstance(solver, pybamm.BaseModel)
 
         # check that adding algebraic variables gives algebraic solver